+++ app/dashboard/(overview)/loading.tsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import DashboardSkeleton from '../../ui/skeletons'
+
+export default function Loading() {
+  return (
+    <DashboardSkeleton />
+  )
+}
